+++ mne/simulation/__init__.py
@@ -6,4 +6,5 @@ from .source import select_source_in_label, simulate_stc, simulate_sparse_stc
 from .source import SourceSimulator
 from .metrics import (source_estimate_quantification, stc_cosine_score,
                       stc_region_localization_error, stc_precision_score,
-                      stc_recall_score, stc_f1_score, stc_roc_auc_score)
+                      stc_recall_score, stc_f1_score, stc_roc_auc_score,
+                      stc_peak_position_error, stc_spacial_deviation)